% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_estado.R
\name{tabela_estado}
\alias{tabela_estado}
\title{Get data about number of psychologist in each Brazilian state}
\usage{
tabela_estado()
}
\value{
A tibble containing two columns: brazilian state and number of psychologists in given state.
}
\description{
Get data about number of psychologist in each Brazilian state
}
