% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_regional.R
\name{tabela_regional}
\alias{tabela_regional}
\title{Get regional data about number of psychologist in Brazil}
\usage{
tabela_regional()
}
\value{
A tibble containing three columns: regional council, brazilian state (initials) and number of psychologists.
}
\description{
Get regional data about number of psychologist in Brazil
}
