% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_sexo.R
\name{tabela_sexo}
\alias{tabela_sexo}
\title{Get data on the number of psychologists of each gender in Brazil by region}
\usage{
tabela_sexo()
}
\value{
A tibble containing four columns: brazilian state, number of female psychologists, number of male psychologists and number of people that did not report their gender.
}
\description{
Get data on the number of psychologists of each gender in Brazil by region
}
